createAsyncThunk:
createAsyncThunk is used to define an asynchronous thunk action creator named getHomePageVideos.
It takes two arguments:
The first argument is a string representing the action type prefix. It's used to automatically generate action types for pending, fulfilled, and rejected actions.
The second argument is an asynchronous function that performs the API request. It receives two parameters: isNext and getState. isNext is a boolean indicating whether to fetch the next page of results, and getState is a function to access the current Redux state.