In the timeSince function, only one of the conditions will be evaluated and returned. Once a condition is met and its corresponding time interval is calculated, the function will return the value and exit. This ensures that only the most appropriate time interval is returned based on the elapsed time since the provided date.
For example, if the elapsed time corresponds to more than one year, the function will return the number of years elapsed. If the elapsed time is less than one year but more than one month, it will return the number of months elapsed. This pattern continues down to the smallest time unit, ensuring that the most relevant time interval is returned.