const items = response.data.items;: This line extracts the items array from the response data obtained from the API call. This array typically contains the items (videos in this case) returned by the YouTube API.
const parsedData = await parseData(items);: This line calls the parseData function to parse the items obtained from the API response. It awaits the completion of this asynchronous operation to get the parsed data.

const nextPageToken = response.data.nextPageToken;: Here, the nextPageToken is extracted from the response data. This token is often provided by APIs to paginate through large datasets. It indicates the token to use in the next request to fetch the next page of results.

dispatch(updateNextPageToken(nextPageToken));: This line dispatches an action (updateNextPageToken) with the nextPageToken as its payload. Dispatching this action will trigger the corresponding reducer to update the nextPageToken in the Redux state.

return { parsedData: [...videos, ...parsedData], nextPageToken };: Finally, this line returns an object containing the parsed data and the nextPageToken. This object will be used as the payload of the fulfilled action dispatched by the createAsyncThunk function. The parsed data will typically be used to update the video list in the Redux state, while the nextPageToken will be stored for future pagination requests.